backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend dla aplikacji sklepowej"
readme = "README.md"
requires-python = ">=3.12.10"

# uv add <package>
# uv sync --all-extras
# uv lock
dependencies = [
    "boto3>=1.40.56",
    "celery>=5.5.3",
    "django>=5.2.7",
    "django-allauth[headless,headless-spec,mfa]>=65.12.1",
    "django-celery-beat>=2.7.0",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-split-settings>=1.3.2",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-camel-case>=1.4.2",
    "djangorestframework-simplejwt>=5.3.0",
    "flower>=2.0.1",
    "pillow>=12.0.0",
    "psycopg2>=2.9.10 ; sys_platform == 'linux'",
    "psycopg2-binary>=2.9.10 ; sys_platform == 'win32'",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "stripe>=13.0.1",
    "rich>=14.2.0",
    "pack-logger[django]",
    "django-countries>=8.2.0",
    "django-phonenumber-field>=8.4.0",
    "phonenumbers>=9.0.25",
    "fido2>=2.1.1",
    "google-auth",
    "requests",
]

# uv add <package> --optional dev
[project.optional-dependencies]
dev = [
    "django-debug-toolbar>=6.2.0",
    "django-extensions>=4.1",
    "django-silk>=5.4.3",
    "django-stubs[compatible-mypy]>=5.2.9",
    "drf-spectacular>=0.29.0",
    "ipython>=9.11.0",
    "ruff>=0.14.1",
]
test = [
    "factory-boy>=3.3.3",
    "faker>=40.13.0",
    "freezegun>=1.5.5",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-django>=4.12.0",
    "pytest-randomly>=4.0.1",
    "pytest-watch>=4.2.0",
    "pytest-xdist>=3.8.0",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.12.9
target-version = "py312"

[tool.ruff.per-file-ignores]
"*/settings/*.py" = ["F401", "F403", "F405"]

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.uv]
link-mode = "copy"

[tool.uv.sources]
pack-logger = { git = "https://github.com/radthenone/pack-logger.git", tag = "v0.1.3", subdirectory = "backend" }

[tool.pyrefly]
search-path = ["src"]
site-package-path = [".venv/Lib/site-packages"]

exclude = [
    "**/__pycache__",
    "**/migrations",
    "**/node_modules",
    ".venv",
    "venv",
    "build",
    "dist",
]

python_version = "3.12"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "core.settings"

[[tool.mypy.overrides]]
module = ["apps.*.models.*", "apps.models.*"]
ignore_errors = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
pythonpath = ["src"]
testpaths = ["src"]
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = "-q --disable-warnings --reuse-db"
